Unit Selection and Barracks
Made the build menu dynamic, items in the menu will be generated at runtime. For now, the structures are static and will show up the same no matter what, but the units are structure-dependant.
Added Barracks. When selected, a Test Unity shows up in the build menu, and spawns at the selected barracks when pressed.
You can now select and move units around. The movement is very fiddly as this is a first pass.